Room DB, DAO, entities, mappers, converters

+++ feature-score/src/main/java/com/suprematic/feature/score/ScoreUiEvent.kt
@@ -4,11 +4,11 @@ import com.suprematic.domain.entities.Team
 
 sealed class ScoreUiEvent(
     val team: Team? = null,
-    val points: Float = 0f
+    val points: Int = 0
 ) {
     object GameInitialized : ScoreUiEvent()
     object GamePaused: ScoreUiEvent()
     object GameFinalized: ScoreUiEvent()
     object EntryUndone: ScoreUiEvent()
-    class PointsScored(team: Team?, points: Float): ScoreUiEvent(team, points)
+    class PointsScored(team: Team?, points: Int): ScoreUiEvent(team, points)
 }
